Decouple tests on routing/basic_test
