+++ packages/ember/tests/routing/basic_test.js
@@ -4,21 +4,14 @@ import {
   Controller,
   RSVP,
   Object as EmberObject,
-  A as emberA,
-  copy
+  A as emberA
 } from 'ember-runtime';
-import {
-  Route,
-  NoneLocation,
-  HistoryLocation
-} from 'ember-routing';
+import { Route } from 'ember-routing';
 import {
   run,
   get,
-  set,
   Mixin,
-  observer,
-  addObserver
+  observer
 } from 'ember-metal';
 import {
   Component,
@@ -46,16 +39,6 @@ function handleURL(assert, path) {
     }, function(reason) {
       assert.ok(false, 'failed to visit:`' + path + '` reason: `' + QUnit.jsDump.parse(reason));
       throw reason;
-    });
-  });
-}
-
-function handleURLAborts(assert, path) {
-  run(() => {
-    router.handleURL(path).then(function() {
-      assert.ok(false, 'url: `' + path + '` was NOT to be handled');
-    }, function(reason) {
-      assert.ok(reason && reason.message === 'TransitionAborted', 'url: `' + path + '` was to be aborted');
     });
   });
 }
@@ -107,1109 +90,6 @@ QUnit.module('Basic Routing', {
     });
   }
 });
-
-QUnit.test('Nested callbacks are not exited when moving to siblings', function(assert) {
-  let done = assert.async();
-  Router.map(function() {
-    this.route('root', { path: '/' }, function() {
-      this.route('special', { path: '/specials/:menu_item_id', resetNamespace: true });
-    });
-  });
-
-  App.RootRoute = Route.extend({
-    model() {
-      rootModel++;
-      return this._super(...arguments);
-    },
-
-    setupController() {
-      rootSetup++;
-    },
-
-    renderTemplate() {
-      rootRender++;
-    },
-
-    serialize: function() {
-      rootSerialize++;
-      return this._super(...arguments);
-    }
-  });
-
-  let currentPath;
-
-  App.ApplicationController = Controller.extend({
-    currentPathDidChange: observer('currentPath', function() {
-      currentPath = get(this, 'currentPath');
-    })
-  });
-
-  let menuItem;
-
-  App.MenuItem = EmberObject.extend();
-  App.MenuItem.reopenClass({
-    find(id) {
-      menuItem = App.MenuItem.create({ id: id });
-      return menuItem;
-    }
-  });
-
-  App.LoadingRoute = Route.extend({
-
-  });
-
-  App.HomeRoute = Route.extend({
-
-  });
-
-  App.SpecialRoute = Route.extend({
-    setupController(controller, model) {
-      set(controller, 'model', model);
-    }
-  });
-
-  setTemplate('root/index', compile(
-    '<h3>Home</h3>'
-  ));
-
-  setTemplate('special', compile(
-    '<p>{{model.id}}</p>'
-  ));
-
-  setTemplate('loading', compile(
-    '<p>LOADING!</p>'
-  ));
-
-  let rootSetup = 0;
-  let rootRender = 0;
-  let rootModel = 0;
-  let rootSerialize = 0;
-
-  bootApplication();
-
-  registry.register('controller:special', Controller.extend());
-
-  assert.equal(getTextOf(rootElement.querySelector('h3')), 'Home', 'The app is now in the initial state');
-  assert.equal(rootSetup, 1, 'The root setup was triggered');
-  assert.equal(rootRender, 1, 'The root render was triggered');
-  assert.equal(rootSerialize, 0, 'The root serialize was not called');
-  assert.equal(rootModel, 1, 'The root model was called');
-
-  router = container.lookup('router:main');
-
-  run(() => {
-    let menuItem = App.MenuItem.create({ id: 1 });
-    run.later(() => RSVP.resolve(menuItem), 1);
-
-    router.transitionTo('special', menuItem).then(function() {
-      assert.equal(rootSetup, 1, 'The root setup was not triggered again');
-      assert.equal(rootRender, 1, 'The root render was not triggered again');
-      assert.equal(rootSerialize, 0, 'The root serialize was not called');
-
-      // TODO: Should this be changed?
-      assert.equal(rootModel, 1, 'The root model was called again');
-
-      assert.deepEqual(router.location.path, '/specials/1');
-      assert.equal(currentPath, 'root.special');
-
-      done();
-    });
-  });
-});
-
-QUnit.test('Events are triggered on the controller if a matching action name is implemented', function(assert) {
-  let done = assert.async();
-
-  Router.map(function() {
-    this.route('home', { path: '/' });
-  });
-
-  let model = { name: 'Tom Dale' };
-  let stateIsNotCalled = true;
-
-  App.HomeRoute = Route.extend({
-    model() {
-      return model;
-    },
-
-    actions: {
-      showStuff() {
-        stateIsNotCalled = false;
-      }
-    }
-  });
-
-  setTemplate('home', compile(
-    '<a {{action \'showStuff\' model}}>{{name}}</a>'
-  ));
-
-  let controller = Controller.extend({
-    actions: {
-      showStuff(context) {
-        assert.ok(stateIsNotCalled, 'an event on the state is not triggered');
-        assert.deepEqual(context, { name: 'Tom Dale' }, 'an event with context is passed');
-        done();
-      }
-    }
-  });
-
-  registry.register('controller:home', controller);
-
-  bootApplication();
-
-  rootElement.querySelector('a').click();
-});
-
-QUnit.test('Events are triggered on the current state when defined in `actions` object', function(assert) {
-  let done = assert.async();
-  Router.map(function() {
-    this.route('home', { path: '/' });
-  });
-
-  let model = { name: 'Tom Dale' };
-
-  App.HomeRoute = Route.extend({
-    model() {
-      return model;
-    },
-
-    actions: {
-      showStuff(obj) {
-        assert.ok(this instanceof App.HomeRoute, 'the handler is an App.HomeRoute');
-        // Using Ember.copy removes any private Ember vars which older IE would be confused by
-        assert.deepEqual(copy(obj, true), { name: 'Tom Dale' }, 'the context is correct');
-        done();
-      }
-    }
-  });
-
-  setTemplate('home', compile(
-    '<a {{action \'showStuff\' model}}>{{model.name}}</a>'
-  ));
-
-  bootApplication();
-
-  rootElement.querySelector('a').click();
-});
-
-QUnit.test('Events defined in `actions` object are triggered on the current state when routes are nested', function(assert) {
-  let done = assert.async();
-
-  Router.map(function() {
-    this.route('root', { path: '/' }, function() {
-      this.route('index', { path: '/' });
-    });
-  });
-
-  let model = { name: 'Tom Dale' };
-
-  App.RootRoute = Route.extend({
-    actions: {
-      showStuff(obj) {
-        assert.ok(this instanceof App.RootRoute, 'the handler is an App.HomeRoute');
-        // Using Ember.copy removes any private Ember vars which older IE would be confused by
-        assert.deepEqual(copy(obj, true), { name: 'Tom Dale' }, 'the context is correct');
-        done();
-      }
-    }
-  });
-
-  App.RootIndexRoute = Route.extend({
-    model() {
-      return model;
-    }
-  });
-
-  setTemplate('root/index', compile(
-    '<a {{action \'showStuff\' model}}>{{model.name}}</a>'
-  ));
-
-  bootApplication();
-
-  rootElement.querySelector('a').click();
-});
-
-QUnit.test('Events can be handled by inherited event handlers', function(assert) {
-  assert.expect(4);
-
-  App.SuperRoute = Route.extend({
-    actions: {
-      foo() {
-        assert.ok(true, 'foo');
-      },
-      bar(msg) {
-        assert.equal(msg, 'HELLO');
-      }
-    }
-  });
-
-  App.RouteMixin = Mixin.create({
-    actions: {
-      bar(msg) {
-        assert.equal(msg, 'HELLO');
-        this._super(msg);
-      }
-    }
-  });
-
-  App.IndexRoute = App.SuperRoute.extend(App.RouteMixin, {
-    actions: {
-      baz() {
-        assert.ok(true, 'baz');
-      }
-    }
-  });
-
-  bootApplication();
-
-  router.send('foo');
-  router.send('bar', 'HELLO');
-  router.send('baz');
-});
-
-QUnit.test('Actions are not triggered on the controller if a matching action name is implemented as a method', function(assert) {
-  let done = assert.async();
-
-  Router.map(function() {
-    this.route('home', { path: '/' });
-  });
-
-  let model = { name: 'Tom Dale' };
-  let stateIsNotCalled = true;
-
-  App.HomeRoute = Route.extend({
-    model() {
-      return model;
-    },
-
-    actions: {
-      showStuff(context) {
-        assert.ok(stateIsNotCalled, 'an event on the state is not triggered');
-        assert.deepEqual(context, { name: 'Tom Dale' }, 'an event with context is passed');
-        done();
-      }
-    }
-  });
-
-  setTemplate('home', compile(
-    '<a {{action \'showStuff\' model}}>{{name}}</a>'
-  ));
-
-  let controller = Controller.extend({
-    showStuff() {
-      stateIsNotCalled = false;
-      assert.ok(stateIsNotCalled, 'an event on the state is not triggered');
-    }
-  });
-
-  registry.register('controller:home', controller);
-
-  bootApplication();
-
-  rootElement.querySelector('a').click();
-});
-
-QUnit.test('actions can be triggered with multiple arguments', function(assert) {
-  let done = assert.async();
-  Router.map(function() {
-    this.route('root', { path: '/' }, function() {
-      this.route('index', { path: '/' });
-    });
-  });
-
-  let model1 = { name: 'Tilde' };
-  let model2 = { name: 'Tom Dale' };
-
-  App.RootRoute = Route.extend({
-    actions: {
-      showStuff(obj1, obj2) {
-        assert.ok(this instanceof App.RootRoute, 'the handler is an App.HomeRoute');
-        // Using Ember.copy removes any private Ember vars which older IE would be confused by
-        assert.deepEqual(copy(obj1, true), { name: 'Tilde' }, 'the first context is correct');
-        assert.deepEqual(copy(obj2, true), { name: 'Tom Dale' }, 'the second context is correct');
-        done();
-      }
-    }
-  });
-
-  App.RootIndexController = Controller.extend({
-    model1: model1,
-    model2: model2
-  });
-
-  setTemplate('root/index', compile(
-    '<a {{action \'showStuff\' model1 model2}}>{{model1.name}}</a>'
-  ));
-
-  bootApplication();
-
-  rootElement.querySelector('a').click();
-});
-
-QUnit.test('transitioning multiple times in a single run loop only sets the URL once', function(assert) {
-  Router.map(function() {
-    this.route('root', { path: '/' });
-    this.route('foo');
-    this.route('bar');
-  });
-
-  bootApplication();
-
-  let urlSetCount = 0;
-
-  router.get('location').setURL = function(path) {
-    urlSetCount++;
-    set(this, 'path', path);
-  };
-
-  assert.equal(urlSetCount, 0);
-
-  run(function() {
-    router.transitionTo('foo');
-    router.transitionTo('bar');
-  });
-
-  assert.equal(urlSetCount, 1);
-  assert.equal(router.get('location').getURL(), '/bar');
-});
-
-QUnit.test('navigating away triggers a url property change', function(assert) {
-  assert.expect(3);
-
-  Router.map(function() {
-    this.route('root', { path: '/' });
-    this.route('foo', { path: '/foo' });
-    this.route('bar', { path: '/bar' });
-  });
-
-  bootApplication();
-
-  run(() => {
-    addObserver(router, 'url', function() {
-      assert.ok(true, 'url change event was fired');
-    });
-  });
-
-  ['foo', 'bar', '/foo'].forEach(destination => run(router, 'transitionTo', destination));
-});
-
-QUnit.test('using replaceWith calls location.replaceURL if available', function(assert) {
-  let setCount = 0;
-  let replaceCount = 0;
-
-  Router.reopen({
-    location: NoneLocation.create({
-      setURL(path) {
-        setCount++;
-        set(this, 'path', path);
-      },
-
-      replaceURL(path) {
-        replaceCount++;
-        set(this, 'path', path);
-      }
-    })
-  });
-
-  Router.map(function() {
-    this.route('root', { path: '/' });
-    this.route('foo');
-  });
-
-  bootApplication();
-
-  assert.equal(setCount, 0);
-  assert.equal(replaceCount, 0);
-
-  run(() => router.replaceWith('foo'));
-
-  assert.equal(setCount, 0, 'should not call setURL');
-  assert.equal(replaceCount, 1, 'should call replaceURL once');
-  assert.equal(router.get('location').getURL(), '/foo');
-});
-
-QUnit.test('using replaceWith calls setURL if location.replaceURL is not defined', function(assert) {
-  let setCount = 0;
-
-  Router.reopen({
-    location: NoneLocation.create({
-      setURL(path) {
-        setCount++;
-        set(this, 'path', path);
-      }
-    })
-  });
-
-  Router.map(function() {
-    this.route('root', { path: '/' });
-    this.route('foo');
-  });
-
-  bootApplication();
-
-  assert.equal(setCount, 0);
-
-  run(() => router.replaceWith('foo'));
-
-  assert.equal(setCount, 1, 'should call setURL once');
-  assert.equal(router.get('location').getURL(), '/foo');
-});
-
-QUnit.test('Route inherits model from parent route', function(assert) {
-  assert.expect(9);
-
-  Router.map(function() {
-    this.route('the_post', { path: '/posts/:post_id' }, function() {
-      this.route('comments');
-
-      this.route('shares', { path: '/shares/:share_id', resetNamespace: true }, function() {
-        this.route('share');
-      });
-    });
-  });
-
-  let post1 = {};
-  let post2 = {};
-  let post3 = {};
-  let share1 = {};
-  let share2 = {};
-  let share3 = {};
-
-  let posts = {
-    1: post1,
-    2: post2,
-    3: post3
-  };
-  let shares = {
-    1: share1,
-    2: share2,
-    3: share3
-  };
-
-  App.ThePostRoute = Route.extend({
-    model(params) {
-      return posts[params.post_id];
-    }
-  });
-
-  App.ThePostCommentsRoute = Route.extend({
-    afterModel(post /*, transition */) {
-      let parent_model = this.modelFor('thePost');
-
-      assert.equal(post, parent_model);
-    }
-  });
-
-  App.SharesRoute = Route.extend({
-    model(params) {
-      return shares[params.share_id];
-    }
-  });
-
-  App.SharesShareRoute = Route.extend({
-    afterModel(share /*, transition */) {
-      let parent_model = this.modelFor('shares');
-
-      assert.equal(share, parent_model);
-    }
-  });
-
-  bootApplication();
-
-  handleURL(assert, '/posts/1/comments');
-  handleURL(assert, '/posts/1/shares/1');
-
-  handleURL(assert, '/posts/2/comments');
-  handleURL(assert, '/posts/2/shares/2');
-
-  handleURL(assert, '/posts/3/comments');
-  handleURL(assert, '/posts/3/shares/3');
-});
-
-QUnit.test('Routes with { resetNamespace: true } inherits model from parent route', function(assert) {
-  assert.expect(6);
-
-  Router.map(function() {
-    this.route('the_post', { path: '/posts/:post_id' }, function() {
-      this.route('comments', { resetNamespace: true }, function() {
-      });
-    });
-  });
-
-  let post1 = {};
-  let post2 = {};
-  let post3 = {};
-
-  let posts = {
-    1: post1,
-    2: post2,
-    3: post3
-  };
-
-  App.ThePostRoute = Route.extend({
-    model(params) {
-      return posts[params.post_id];
-    }
-  });
-
-  App.CommentsRoute = Route.extend({
-    afterModel(post /*, transition */) {
-      let parent_model = this.modelFor('thePost');
-
-      assert.equal(post, parent_model);
-    }
-  });
-
-  bootApplication();
-
-  handleURL(assert, '/posts/1/comments');
-  handleURL(assert, '/posts/2/comments');
-  handleURL(assert, '/posts/3/comments');
-});
-
-QUnit.test('It is possible to get the model from a parent route', function(assert) {
-  assert.expect(9);
-
-  Router.map(function() {
-    this.route('the_post', { path: '/posts/:post_id' }, function() {
-      this.route('comments', { resetNamespace: true });
-    });
-  });
-
-  let post1 = {};
-  let post2 = {};
-  let post3 = {};
-  let currentPost;
-
-  let posts = {
-    1: post1,
-    2: post2,
-    3: post3
-  };
-
-  App.ThePostRoute = Route.extend({
-    model(params) {
-      return posts[params.post_id];
-    }
-  });
-
-  App.CommentsRoute = Route.extend({
-    model() {
-      // Allow both underscore / camelCase format.
-      assert.equal(this.modelFor('thePost'), currentPost);
-      assert.equal(this.modelFor('the_post'), currentPost);
-    }
-  });
-
-  bootApplication();
-
-  currentPost = post1;
-  handleURL(assert, '/posts/1/comments');
-
-  currentPost = post2;
-  handleURL(assert, '/posts/2/comments');
-
-  currentPost = post3;
-  handleURL(assert, '/posts/3/comments');
-});
-
-QUnit.test('A redirection hook is provided', function(assert) {
-  Router.map(function() {
-    this.route('choose', { path: '/' });
-    this.route('home');
-  });
-
-  let chooseFollowed = 0;
-  let destination;
-
-  App.ChooseRoute = Route.extend({
-    redirect() {
-      if (destination) {
-        this.transitionTo(destination);
-      }
-    },
-
-    setupController() {
-      chooseFollowed++;
-    }
-  });
-
-  destination = 'home';
-
-  bootApplication();
-
-  assert.equal(chooseFollowed, 0, 'The choose route wasn\'t entered since a transition occurred');
-  assert.equal(rootElement.querySelectorAll('h3.hours').length, 1, 'The home template was rendered');
-  assert.equal(getOwner(router).lookup('controller:application').get('currentPath'), 'home');
-});
-
-QUnit.test('Redirecting from the middle of a route aborts the remainder of the routes', function(assert) {
-  assert.expect(3);
-
-  Router.map(function() {
-    this.route('home');
-    this.route('foo', function() {
-      this.route('bar', { resetNamespace: true }, function() {
-        this.route('baz');
-      });
-    });
-  });
-
-  App.BarRoute = Route.extend({
-    redirect() {
-      this.transitionTo('home');
-    },
-    setupController() {
-      assert.ok(false, 'Should transition before setupController');
-    }
-  });
-
-  App.BarBazRoute = Route.extend({
-    enter() {
-      assert.ok(false, 'Should abort transition getting to next route');
-    }
-  });
-
-  bootApplication();
-
-  handleURLAborts(assert, '/foo/bar/baz');
-
-  assert.equal(getOwner(router).lookup('controller:application').get('currentPath'), 'home');
-  assert.equal(router.get('location').getURL(), '/home');
-});
-
-QUnit.test('Redirecting to the current target in the middle of a route does not abort initial routing', function(assert) {
-  assert.expect(5);
-
-  Router.map(function() {
-    this.route('home');
-    this.route('foo', function() {
-      this.route('bar', { resetNamespace: true }, function() {
-        this.route('baz');
-      });
-    });
-  });
-
-  let successCount = 0;
-  App.BarRoute = Route.extend({
-    redirect() {
-      this.transitionTo('bar.baz').then(function() {
-        successCount++;
-      });
-    },
-
-    setupController() {
-      assert.ok(true, 'Should still invoke bar\'s setupController');
-    }
-  });
-
-  App.BarBazRoute = Route.extend({
-    setupController() {
-      assert.ok(true, 'Should still invoke bar.baz\'s setupController');
-    }
-  });
-
-  bootApplication();
-
-  handleURL(assert, '/foo/bar/baz');
-
-  assert.equal(getOwner(router).lookup('controller:application').get('currentPath'), 'foo.bar.baz');
-  assert.equal(successCount, 1, 'transitionTo success handler was called once');
-});
-
-QUnit.test('Redirecting to the current target with a different context aborts the remainder of the routes', function(assert) {
-  assert.expect(4);
-
-  Router.map(function() {
-    this.route('home');
-    this.route('foo', function() {
-      this.route('bar', { path: 'bar/:id', resetNamespace: true }, function() {
-        this.route('baz');
-      });
-    });
-  });
-
-  let model = { id: 2 };
-
-  let count = 0;
-
-  App.BarRoute = Route.extend({
-    afterModel() {
-      if (count++ > 10) {
-        assert.ok(false, 'infinite loop');
-      } else {
-        this.transitionTo('bar.baz', model);
-      }
-    }
-  });
-
-  App.BarBazRoute = Route.extend({
-    setupController() {
-      assert.ok(true, 'Should still invoke setupController');
-    }
-  });
-
-  bootApplication();
-
-  handleURLAborts(assert, '/foo/bar/1/baz');
-
-  assert.equal(getOwner(router).lookup('controller:application').get('currentPath'), 'foo.bar.baz');
-  assert.equal(router.get('location').getURL(), '/foo/bar/2/baz');
-});
-
-QUnit.test('Transitioning from a parent event does not prevent currentPath from being set', function(assert) {
-  Router.map(function() {
-    this.route('foo', function() {
-      this.route('bar', { resetNamespace: true }, function() {
-        this.route('baz');
-      });
-      this.route('qux');
-    });
-  });
-
-  App.FooRoute = Route.extend({
-    actions: {
-      goToQux() {
-        this.transitionTo('foo.qux');
-      }
-    }
-  });
-
-  bootApplication();
-
-  let applicationController = getOwner(router).lookup('controller:application');
-
-  handleURL(assert, '/foo/bar/baz');
-
-  assert.equal(applicationController.get('currentPath'), 'foo.bar.baz');
-
-  run(() => router.send('goToQux'));
-
-  assert.equal(applicationController.get('currentPath'), 'foo.qux');
-  assert.equal(router.get('location').getURL(), '/foo/qux');
-});
-
-QUnit.test('Generated names can be customized when providing routes with dot notation', function(assert) {
-  assert.expect(4);
-
-  setTemplate('index', compile('<div>Index</div>'));
-  setTemplate('application', compile('<h1>Home</h1><div class=\'main\'>{{outlet}}</div>'));
-  setTemplate('foo', compile('<div class=\'middle\'>{{outlet}}</div>'));
-  setTemplate('bar', compile('<div class=\'bottom\'>{{outlet}}</div>'));
-  setTemplate('bar/baz', compile('<p>{{name}}Bottom!</p>'));
-
-  Router.map(function() {
-    this.route('foo', { path: '/top' }, function() {
-      this.route('bar', { path: '/middle', resetNamespace: true }, function() {
-        this.route('baz', { path: '/bottom' });
-      });
-    });
-  });
-
-  App.FooRoute = Route.extend({
-    renderTemplate() {
-      assert.ok(true, 'FooBarRoute was called');
-      return this._super(...arguments);
-    }
-  });
-
-  App.BarBazRoute = Route.extend({
-    renderTemplate() {
-      assert.ok(true, 'BarBazRoute was called');
-      return this._super(...arguments);
-    }
-  });
-
-  App.BarController = Controller.extend({
-    name: 'Bar'
-  });
-
-  App.BarBazController = Controller.extend({
-    name: 'BarBaz'
-  });
-
-  bootApplication();
-
-  handleURL(assert, '/top/middle/bottom');
-
-  assert.equal(getTextOf(rootElement.querySelector('.main .middle .bottom p')), 'BarBazBottom!', 'The templates were rendered into their appropriate parents');
-});
-
-QUnit.test('Child routes render into their parent route\'s template by default', function(assert) {
-  setTemplate('index', compile('<div>Index</div>'));
-  setTemplate('application', compile('<h1>Home</h1><div class=\'main\'>{{outlet}}</div>'));
-  setTemplate('top', compile('<div class=\'middle\'>{{outlet}}</div>'));
-  setTemplate('middle', compile('<div class=\'bottom\'>{{outlet}}</div>'));
-  setTemplate('middle/bottom', compile('<p>Bottom!</p>'));
-
-  Router.map(function() {
-    this.route('top', function() {
-      this.route('middle', { resetNamespace: true }, function() {
-        this.route('bottom');
-      });
-    });
-  });
-
-  bootApplication();
-
-  handleURL(assert, '/top/middle/bottom');
-
-  assert.equal(getTextOf(rootElement.querySelector('.main .middle .bottom p')), 'Bottom!', 'The templates were rendered into their appropriate parents');
-});
-
-QUnit.test('Child routes render into specified template', function(assert) {
-  setTemplate('index', compile('<div>Index</div>'));
-  setTemplate('application', compile('<h1>Home</h1><div class=\'main\'>{{outlet}}</div>'));
-  setTemplate('top', compile('<div class=\'middle\'>{{outlet}}</div>'));
-  setTemplate('middle', compile('<div class=\'bottom\'>{{outlet}}</div>'));
-  setTemplate('middle/bottom', compile('<p>Bottom!</p>'));
-
-  Router.map(function() {
-    this.route('top', function() {
-      this.route('middle', { resetNamespace: true }, function() {
-        this.route('bottom');
-      });
-    });
-  });
-
-  App.MiddleBottomRoute = Route.extend({
-    renderTemplate() {
-      this.render('middle/bottom', { into: 'top' });
-    }
-  });
-
-  bootApplication();
-
-  handleURL(assert, '/top/middle/bottom');
-
-  assert.equal(rootElement.querySelectorAll('.main .middle .bottom p').length, 0, 'should not render into the middle template');
-  assert.equal(getTextOf(rootElement.querySelector('.main .middle > p')), 'Bottom!', 'The template was rendered into the top template');
-});
-
-QUnit.test('Rendering into specified template with slash notation', function(assert) {
-  setTemplate('person/profile', compile('profile {{outlet}}'));
-  setTemplate('person/details', compile('details!'));
-
-  Router.map(function() {
-    this.route('home', { path: '/' });
-  });
-
-  App.HomeRoute = Route.extend({
-    renderTemplate() {
-      this.render('person/profile');
-      this.render('person/details', { into: 'person/profile' });
-    }
-  });
-
-  bootApplication();
-
-  assert.equal(rootElement.textContent.trim(), 'profile details!', 'The templates were rendered');
-});
-
-QUnit.test('Parent route context change', function(assert) {
-  let editCount = 0;
-  let editedPostIds = emberA();
-
-  setTemplate('application', compile('{{outlet}}'));
-  setTemplate('posts', compile('{{outlet}}'));
-  setTemplate('post', compile('{{outlet}}'));
-  setTemplate('post/index', compile('showing'));
-  setTemplate('post/edit', compile('editing'));
-
-  Router.map(function() {
-    this.route('posts', function() {
-      this.route('post', { path: '/:postId', resetNamespace: true }, function() {
-        this.route('edit');
-      });
-    });
-  });
-
-  App.PostsRoute = Route.extend({
-    actions: {
-      showPost(context) {
-        this.transitionTo('post', context);
-      }
-    }
-  });
-
-  App.PostRoute = Route.extend({
-    model(params) {
-      return { id: params.postId };
-    },
-
-    serialize(model) {
-      return { postId: model.id };
-    },
-
-    actions: {
-      editPost() {
-        this.transitionTo('post.edit');
-      }
-    }
-  });
-
-  App.PostEditRoute = Route.extend({
-    model() {
-      let postId = this.modelFor('post').id;
-      editedPostIds.push(postId);
-      return null;
-    },
-    setup() {
-      this._super(...arguments);
-      editCount++;
-    }
-  });
-
-  bootApplication();
-
-  handleURL(assert, '/posts/1');
-
-  run(() => router.send('editPost'));
-  run(() => router.send('showPost', { id: '2' }));
-  run(() => router.send('editPost'));
-
-  assert.equal(editCount, 2, 'set up the edit route twice without failure');
-  assert.deepEqual(editedPostIds, ['1', '2'], 'modelFor posts.post returns the right context');
-});
-
-QUnit.test('Router accounts for rootURL on page load when using history location', function(assert) {
-  let rootURL = window.location.pathname + '/app';
-  let postsTemplateRendered = false;
-  let setHistory, HistoryTestLocation;
-
-  setHistory = function(obj, path) {
-    obj.set('history', { state: { path: path } });
-  };
-
-  // Create new implementation that extends HistoryLocation
-  // and set current location to rootURL + '/posts'
-  HistoryTestLocation = HistoryLocation.extend({
-    initState() {
-      let path = rootURL + '/posts';
-
-      setHistory(this, path);
-      this.set('location', {
-        pathname: path,
-        href: 'http://localhost/' + path
-      });
-    },
-
-    replaceState(path) {
-      setHistory(this, path);
-    },
-
-    pushState(path) {
-      setHistory(this, path);
-    }
-  });
-
-
-  registry.register('location:historyTest', HistoryTestLocation);
-
-  Router.reopen({
-    location: 'historyTest',
-    rootURL: rootURL
-  });
-
-  Router.map(function() {
-    this.route('posts', { path: '/posts' });
-  });
-
-  App.PostsRoute = Route.extend({
-    model() {},
-    renderTemplate() {
-      postsTemplateRendered = true;
-    }
-  });
-
-  bootApplication();
-
-  assert.ok(postsTemplateRendered, 'Posts route successfully stripped from rootURL');
-});
-
-QUnit.test('The rootURL is passed properly to the location implementation', function(assert) {
-  assert.expect(1);
-  let rootURL = '/blahzorz';
-  let HistoryTestLocation;
-
-  HistoryTestLocation = HistoryLocation.extend({
-    rootURL: 'this is not the URL you are looking for',
-    initState() {
-      assert.equal(this.get('rootURL'), rootURL);
-    }
-  });
-
-  registry.register('location:history-test', HistoryTestLocation);
-
-  Router.reopen({
-    location: 'history-test',
-    rootURL: rootURL,
-    // if we transition in this test we will receive failures
-    // if the tests are run from a static file
-    _doURLTransition() { }
-  });
-
-  bootApplication();
-});
-
-
-QUnit.test('Only use route rendered into main outlet for default into property on child', function(assert) {
-  setTemplate('application', compile('{{outlet \'menu\'}}{{outlet}}'));
-  setTemplate('posts', compile('{{outlet}}'));
-  setTemplate('posts/index', compile('<p class="posts-index">postsIndex</p>'));
-  setTemplate('posts/menu', compile('<div class="posts-menu">postsMenu</div>'));
-
-  Router.map(function() {
-    this.route('posts', function() {});
-  });
-
-  App.PostsRoute = Route.extend({
-    renderTemplate() {
-      this.render();
-      this.render('posts/menu', {
-        into: 'application',
-        outlet: 'menu'
-      });
-    }
-  });
-
-  bootApplication();
-
-  handleURL(assert, '/posts');
-
-  assert.equal(getTextOf(rootElement.querySelector('div.posts-menu')), 'postsMenu', 'The posts/menu template was rendered');
-  assert.equal(getTextOf(rootElement.querySelector('p.posts-index')), 'postsIndex', 'The posts/index template was rendered');
-});
-
-QUnit.test('Generating a URL should not affect currentModel', function(assert) {
-  Router.map(function() {
-    this.route('post', { path: '/posts/:post_id' });
-  });
-
-  let posts = {
-    1: { id: 1 },
-    2: { id: 2 }
-  };
-
-  App.PostRoute = Route.extend({
-    model(params) {
-      return posts[params.post_id];
-    }
-  });
-
-  bootApplication();
-
-  handleURL(assert, '/posts/1');
-
-  let route = container.lookup('route:post');
-  assert.equal(route.modelFor('post'), posts[1]);
-
-  let url = router.generate('post', posts[2]);
-  assert.equal(url, '/posts/2');
-
-  assert.equal(route.modelFor('post'), posts[1]);
-});
-
 
 QUnit.test('Generated route should be an instance of App.Route if provided', function(assert) {
   let generatedRoute;
@@ -1478,7 +358,6 @@ QUnit.test('Route should tear down multiple outlets', function(assert) {
   assert.equal(rootElement.querySelector('p.posts-index'), null, 'The posts/index template was removed');
   assert.equal(rootElement.querySelector('div.posts-footer'), null, 'The posts/footer template was removed');
 });
-
 
 QUnit.test('Route will assert if you try to explicitly render {into: ...} a missing template', function() {
   expectDeprecation(/Rendering into a {{render}} helper that resolves to an {{outlet}} is deprecated./);
@@ -2323,7 +1202,6 @@ QUnit.test('Exception during initialization of non-initial route is not swallowe
   assert.throws(() => run(router, 'transitionTo', 'boom'), /\bboom\b/);
 });
 
-
 QUnit.test('Exception during load of non-initial route is not swallowed', function(assert) {
   Router.map(function() {
     this.route('boom');
@@ -2473,7 +1351,6 @@ QUnit.test('Can render into a named outlet at the top level, with empty main out
 
   assert.equal(rootElement.textContent.trim(), 'A--B-Hello world-C', 'initial render');
 });
-
 
 QUnit.test('Can render into a named outlet at the top level, later', function(assert) {
   setTemplate('application', compile('A-{{outlet}}-B-{{outlet "other"}}-C'));
